services/mr-m/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Pipeline for extracting, cleaning, and chunking content from frontend and PDF sources, preparing it for embeddings/ llm input."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-limiter>=3.12",
    "fuzzywuzzy>=0.18.0",
    "httpie>=3.2.4",
    "openai>=1.95.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "requests>=2.32.4",
    "spacy>=3.7.2,<3.8.0",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
    "numpy==1.26.4",
    "pillow>=11.3.0",
    "tiktoken>=0.9.0",
    "filelock>=3.18.0",
    "opencv-python>=4.11.0.86",
    "redis>=5.0.0,<6.0.0",
    "werkzeug>=3.0",
    "boto3>=1.35.0" 
]
